requirements.txt
allure-pytest
pytest
pytest-selenium
selenium
requests